+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Update Log
 
 - [Update Log](#update-log)
+  - [gh 0.0.15 | Cool title effect](#gh-0015--cool-title-effect)
   - [gh 0.0.14 | Social links redirects](#gh-0014--social-links-redirects)
   - [gh 0.0.13 | Improved experience on phones](#gh-0013--improved-experience-on-phones)
   - [gh 0.0.12 | Console cleans after last command](#gh-0012--console-cleans-after-last-command)
@@ -17,6 +18,7 @@
   - [gh 0.0.1 | Surce Code](#gh-001--surce-code)
 
 
+## gh 0.0.15 | Cool title effect
 ## gh 0.0.14 | Social links redirects
 ## gh 0.0.13 | Improved experience on phones
 ## gh 0.0.12 | Console cleans after last command